refactor: rename NE packages and export Resource/DataSource
