+++ internal/resources/networkedge/account/data_source_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package account
 
 import (
 	"fmt"
