+++ internal/resources/networkedge/acl_template/resource_test.go
@@ -1,4 +1,4 @@
-package equinix
+package acl_template
 
 import (
 	"testing"
