+++ internal/resources/networkedge/bgp/resource_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package bgp
 
 import (
 	"fmt"
