+++ internal/resources/networkedge/device/data_source_test.go
@@ -1,4 +1,4 @@
-package equinix
+package device
 
 import (
 	"reflect"
