+++ internal/resources/networkedge/device/resource_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package device
 
 import (
 	"context"
