+++ internal/resources/networkedge/device_link/resource_test.go
@@ -1,4 +1,4 @@
-package equinix
+package device_link
 
 import (
 	"testing"
