+++ internal/resources/networkedge/device_software/data_source.go
@@ -1,4 +1,4 @@
-package equinix
+package device_software
 
 import (
 	"context"
@@ -45,7 +45,7 @@ var networkDeviceSoftwareDescriptions = map[string]string{
 
 const networkDeviceSoftwareDateLayout = "2006-01-02"
 
-func dataSourceNetworkDeviceSoftware() *schema.Resource {
+func DataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceNetworkDeviceSoftwareRead,
 		Description: "Use this data source to get Equinix Network Edge device software details for a given device type.",
