+++ internal/resources/networkedge/device_software/data_source_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package device_software
 
 import (
 	"fmt"
