+++ internal/resources/networkedge/device_type/data_source.go
@@ -1,4 +1,4 @@
-package equinix
+package device_type
 
 import (
 	"context"
@@ -34,7 +34,7 @@ var networkDeviceTypeDescriptions = map[string]string{
 	"MetroCodes":  "List of metro codes where device type has to be available",
 }
 
-func dataSourceNetworkDeviceType() *schema.Resource {
+func DataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceNetworkDeviceTypeRead,
 		Description: "Use this data source to get Equinix Network Edge device type details",
