+++ internal/resources/networkedge/device_type/data_source_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package device_type
 
 import (
 	"fmt"
