+++ internal/resources/networkedge/file/resource_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package file
 
 import (
 	"fmt"
