+++ internal/resources/networkedge/platform/data_source_acc_test.go
@@ -1,4 +1,4 @@
-package equinix
+package platform
 
 import (
 	"fmt"
