+++ internal/resources/networkedge/ssh_key/resource_test.go
@@ -1,4 +1,4 @@
-package equinix
+package ssh_key
 
 import (
 	"testing"
